refactor: bug fixes, add logging
