--- brbot/.idea/.name
@@ -1,1 +1,0 @@
-cog.py
