+++ brbot/Core/anilist.py
@@ -1,5 +1,4 @@
 import httpx
-import asyncio
 
 
 def anilist_id_from_url(url: str, is_character: bool = False) -> int | None:
@@ -17,15 +16,6 @@ def anilist_id_from_url(url: str, is_character: bool = False) -> int | None:
         return int(url[-3])
 
     return None
-
-
-def username_from_url(url: str) -> str | None:
-    url = url.split("/")
-    url = tuple(filter(None, url))
-    if not url:
-        return None
-
-    return url[-1]
 
 
 def query_media(*, media_id: int):
@@ -63,146 +53,6 @@ def query_media(*, media_id: int):
         return None
 
     return response.json()["data"]["Media"]
-
-
-def query_character(*, character_id: int):
-    query = """
-    query Character($characterId: Int) {
-      Character(id: $characterId) {
-        image {
-          medium
-        }
-        name {
-          full
-        }
-        siteUrl
-      }
-    }
-    """
-    variables = {
-        "characterId": character_id
-    }
-    response = httpx.post(
-        url="https://graphql.anilist.co",
-        json={"query": query, "variables": variables}
-    )
-    if response.status_code != 200:
-        return None
-
-    return response.json()["data"]["Character"]
-
-
-async def query_media_list_recs(*, user_id: int, manga: bool = False):
-    media_type = "manga" if manga else "anime"
-    query = f"""
-    query User($userId: Int) {{
-      User(id: $userId) {{
-        statistics {{
-          {media_type} {{
-            count
-            meanScore
-            standardDeviation
-            genres {{
-              count
-              genre
-              meanScore
-            }}
-          }}
-        }}
-      }}
-    }}
-    """
-    variables = {
-        "userId": user_id,
-    }
-
-    try:
-        response = httpx.post(
-            url="https://graphql.anilist.co",
-            json={"query": query, "variables": variables},
-        )
-    except httpx.ReadTimeout:
-        return None
-
-    if response.status_code != 200:
-        return None
-    watched_count = response.json()["data"]["User"]["statistics"][media_type]["count"]
-    user_statistics = response.json()["data"]["User"]["statistics"][media_type]
-
-    if not watched_count:
-        return None
-
-    query = """
-    query MediaListCollection($userId: Int, $type: MediaType, $statusNotIn: [MediaListStatus], $sort: [RecommendationSort], $perPage: Int, $perChunk: Int, $chunk: Int) {
-      MediaListCollection(userId: $userId, type: $type, status_not_in: $statusNotIn, perChunk: $perChunk, chunk: $chunk) {
-        lists {
-          entries {
-            score
-            media {
-              id
-              popularity
-              recommendations(sort: $sort, perPage: $perPage) {
-                nodes {
-                  rating
-                  mediaRecommendation {
-                    id
-                    genres
-                    meanScore
-                    popularity
-                    relations {
-                      edges {
-                        relationType
-                      }
-                      nodes {
-                        id
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    """
-    chunk_size = 100
-
-    async def query_list_recommendations(session: httpx.AsyncClient, chunk):
-        req_vars = {
-            "userId": user_id,
-            "type": "MANGA" if manga else "ANIME",
-            "statusNotIn": "PLANNING",
-            "perPage": 8,
-            "sort": "RATING_DESC",
-            "perChunk": chunk_size,
-            "chunk": chunk
-        }
-        data = await session.post(
-            url="https://graphql.anilist.co",
-            json={"query": query, "variables": req_vars},
-            timeout=10
-        )
-        return data
-
-    tasks: list = []
-
-    async with httpx.AsyncClient() as client:
-        for i in range(1, watched_count // chunk_size + 2):
-            tasks.append(query_list_recommendations(client, i))
-
-        raw_list_data = await asyncio.gather(*tasks)
-
-    full_rec_list: list = []
-    for data_chunk in raw_list_data:
-        if data_chunk.status_code != 200:
-            continue
-        data_chunk = data_chunk.json()["data"]["MediaListCollection"]["lists"]
-        for anime_list in data_chunk:
-            anime_list = anime_list["entries"]
-            full_rec_list += anime_list
-
-    return user_statistics, full_rec_list
 
 
 def query_user_id(username: str) -> int | None:
@@ -300,22 +150,28 @@ async def query_user_genres(anilist_user_id: int) -> str | None:
     return least_watched_genre
 
 
-def find_anilist_changes(start_anilist: list[dict], end_anilist: list[dict]) -> list[dict]:
+def query_character(*, character_id: int):
+    query = """
+    query Character($characterId: Int) {
+      Character(id: $characterId) {
+        image {
+          medium
+        }
+        name {
+          full
+        }
+        siteUrl
+      }
+    }
+    """
+    variables = {
+        "characterId": character_id
+    }
+    response = httpx.post(
+        url="https://graphql.anilist.co",
+        json={"query": query, "variables": variables}
+    )
+    if response.status_code != 200:
+        return None
 
-    anilist_changes = []
-    for end_anime in end_anilist:
-        start_anime = next(
-            (start_anime for start_anime in start_anilist if start_anime["mediaId"] == end_anime["mediaId"]), None
-        )
-        if start_anime == end_anime:  # Skip if the show is the same at the beginning and end of game
-            continue
-
-        if not start_anime:  # Show was not on player's anilist when the game started
-            anilist_changes.append(end_anime)
-        else:
-            episode_changes = end_anime["progress"] - start_anime["progress"]
-            anilist_changes.append(
-                {"mediaId": end_anime["mediaId"], "status": end_anime["status"], "progress": episode_changes}
-            )
-
-    return anilist_changes
+    return response.json()["data"]["Character"]
