+++ brbot/Features/Bingo/cog.py
@@ -197,9 +197,10 @@ class BingoCog(commands.GroupCog, name='bingo'):
             await ctx.followup.send(content="🟥")
 
         game.update_game_after_shot(ctx=ctx, shot=shot, player_idx=sender_idx, hit_tile=hit_tile)
+        return False
 
     @shot.autocomplete("tag")
-    async def autocomplete(self, ctx: Interaction, current: str):
+    async def autocomplete(self, _: Interaction, current: str):
         tags = bi.bingo_tags + bi.character_tags + bi.season_tags + bi.episode_tags
         tags = [tag for tag in tags if current.lower() in tag.lower()]
         choices = list(map(bu.autocomplete_filter, tags))
